Add NWN.Parsers.ExoBuildBuffer, NWN.Parsers.ExoParseBuffer.
These classes mirror the native classes of the same name, in the
nwn2dev project, but provide a managed-accessible interface that
is designed for use from C# code.

Basic network level types are also added in under the NWN
namespace, for use by managed code.  Armor and armor accessory
classes were not added at this stage as they are not currently
needed for on the wire messaging.

Additionally, add in the initial project for the master server
program (no contents yet).